bot/test_deepgram.py
#!/usr/bin/env python3
"""
Test Deepgram API integration.
Tests transcription with a real audio file.
"""

import os
import sys
import subprocess
from pathlib import Path
from dotenv import load_dotenv
from deepgram import DeepgramClient, PrerecordedOptions, FileSource

# Load environment
load_dotenv()

def create_test_audio():
    """Create a test WAV file with synthesized speech."""
    test_wav = "test_audio.wav"

    # Create test audio with tone (500Hz sine wave for 2 seconds)
    # This simulates voice better than silence for testing
    result = subprocess.run([
        'ffmpeg', '-y', '-f', 'lavfi',
        '-i', 'sine=frequency=500:duration=2:sample_rate=16000',
        '-acodec', 'pcm_s16le', test_wav
    ], capture_output=True, text=True)

    if result.returncode != 0:
        print(f"ERROR: Failed to create test audio: {result.stderr}")
        return None

    return test_wav

def test_deepgram_api():
    """Test Deepgram API with Nova-3 multilingual model."""
    print("=" * 60)
    print("Testing Deepgram API Integration")
    print("=" * 60)

    # Check API key
    api_key = os.getenv('DEEPGRAM_API_KEY')
    if not api_key:
        print("❌ DEEPGRAM_API_KEY not found in .env")
        return False

    print(f"✓ API Key loaded: {api_key[:10]}...{api_key[-5:]}")

    # Initialize client
    try:
        deepgram = DeepgramClient(api_key)
        print("✓ Deepgram client initialized")
    except Exception as e:
        print(f"❌ Failed to initialize client: {e}")
        return False

    # Create test audio
    print("\nCreating test audio file...")
    test_file = create_test_audio()
    if not test_file:
        return False

    file_size = os.path.getsize(test_file)
    print(f"✓ Test audio created: {test_file} ({file_size} bytes)")

    # Test transcription
    print("\nTesting transcription with Nova-3 multilingual...")
    try:
        with open(test_file, 'rb') as audio:
            # Prepare source (just use dict, FileSource is a type hint)
            source = {'buffer': audio, 'mimetype': 'audio/wav'}

            # Configure options (same as bot.py)
            options = PrerecordedOptions(
                model='nova-3',  # Nova-3 multilingual model
                smart_format=True,
                language='multi'  # Multilingual auto-detection
            )

            print(f"  Model: nova-3")
            print(f"  Language: multi (auto-detect)")
            print(f"  Smart format: True")

            # Call API
            response = deepgram.listen.prerecorded.v('1').transcribe_file(source, options)

            # Extract results
            transcript = response.results.channels[0].alternatives[0].transcript
            confidence = response.results.channels[0].alternatives[0].confidence

            # Get metadata
            metadata = response.metadata
            request_id = metadata.request_id if hasattr(metadata, 'request_id') else metadata.get('request_id', 'unknown')
            model_info = metadata.model_info if hasattr(metadata, 'model_info') else metadata.get('model_info', {})

            print("\n" + "=" * 60)
            print("TRANSCRIPTION RESULTS")
            print("=" * 60)
            print(f"Transcript: '{transcript}' (empty = silence)")
            print(f"Confidence: {confidence:.4f}")
            print(f"Request ID: {request_id}")

            # Extract model info
            if isinstance(model_info, dict):
                model_name = model_info.get('name', 'unknown')
                model_version = model_info.get('version', 'unknown')
            else:
                model_name = getattr(model_info, 'name', 'unknown')
                model_version = getattr(model_info, 'version', 'unknown')

            print(f"Model: {model_name}")
            print(f"Version: {model_version}")

            # Validate configuration (not response, since we sent silence)
            print("\n" + "=" * 60)
            print("VALIDATION")
            print("=" * 60)
            print(f"✓ Requested model: nova-3")
            print(f"✓ Requested language: multi (multilingual)")
            print(f"✓ Smart format: enabled")
            print(f"✓ API connection: successful")
            print(f"✓ Transcription: working (empty output expected for silence)")

            print("\n✓ Transcription successful!")
            print("=" * 60)

            return True

    except Exception as e:
        print(f"\n❌ Transcription failed: {e}")
        import traceback
        traceback.print_exc()
        return False

    finally:
        # Cleanup
        if test_file and os.path.exists(test_file):
            os.remove(test_file)
            print(f"\n✓ Cleaned up test file: {test_file}")

def main():
    success = test_deepgram_api()

    if success:
        print("\n✅ All tests passed!")
        print("Deepgram integration is working correctly.")
        return 0
    else:
        print("\n❌ Tests failed!")
        print("Please check the errors above.")
        return 1

if __name__ == '__main__':
    sys.exit(main())
